[PB-1525] edit handleWebsiteLanguage service
